Crate hlist

This crate provides types Nil and Cons<H, T>, which together allow for creating lists consisting of multiple types. The types in the list are present in the type of the list, so that Cons<i32, Cons<i64, Nil>> contains an i32 and an i64. If type T is present exactly once in an HList, it is usually possible to allow the compiler to find the T in the HList by using the Find trait.

Structs

Cons
An HList with H at position 0, and T as the rest of the list.

Nil
The empty HList.

There
Used as an index into an HList.

Enums

Here
Used as an index into an HList.

Traits

Find
Find<T, I> is implemented for an HList if index I of the HList is a T

HList
A marker trait that Nil and Cons<H, T> satisfies. Not currently used to enforce proper hlists, although this may change. Provides the push() method